Revert "Application routes should be engine routes"
This reverts commit 265600ec4e6d2f39996a585d159067bc348d91e7.

@joneslee85 the routes should remain Application routes otherwise they are not available to the Host Application.

The change to engine routes breaks the extension resulting in undefined_method email_to_friend_url errors rendering the override.

I think it makes more sense to leave the routes as application routes in order to avoid having users mount the the engine in the host applications route configuration, unless you know a way we could do that as a railtie?